// lims-frontend/src/types/apps/reagenTypes.ts
export type ReagenType = {
    createdAt: string;
    updatedAt: string | null;
    createdBy: string;
    updatedBy: string | null;
    id: number | null;
    no_cas: string | null;
    reagen: string |null;
    stock: number | null;
    min_stock: number | null;
    keterangan: string | null;
    form_referensi_id: number | null;
    satuan_referensi_id: number | null;
    uuid: string;
}